faker/pyproject.toml
[tool.poetry]
name = "faker"
version = "0.1.0"
description = ""
authors = ["Marlon Rozindo <marlon.saura@owshq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
numpy = "^1.24.3"
pandas = "^2.0.1"
sqlalchemy = "^2.0.12"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
isort = "^5.12.0"
black = "^23.3.0"
taskipy = "^1.10.4"


[tool.poetry.group.doc.dependencies]
mkdocs = "^1.4.3"
mkdocstrings = "^0.21.2"
mkdocstrings-python = "^0.10.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
black = "^23.3.0"
isort = "^5.12.0"
taskipy = "^1.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 88

[tool.taskipy.tasks]
lint = "black --check --diff . && isort --check --diff ."
forcelint = "black . && isort ."
docs = "mkdocs serve"
test = "pytest -s -x -vv --cov=src"
pre_test = "task lint"
post_test = "coverage html"